requirements.txt
flask==2.3.3
werkzeug==2.3.7
yt-dlp
APScheduler==3.9.1
python-dotenv==0.19.0
Flask-SQLAlchemy==3.1.1
mutagen==1.47.0
Flask-SocketIO>=5.3.0
eventlet>=0.33.0
gunicorn>=21.2.0
pygame
Flask-CORS
Flask-WTF